refactor(agent): split providers/index.ts into types/catalog/model-overlay
Tier 1 item from docs/reports/REPORT_LARGE_FILE_MODULARIZATION_SCAN_2026_07_22.md.
Bundled type contracts, a static 9-CLI config table, and a dynamic
model-catalog overlay (version-comparison logic unrelated to what a provider
is) in one 633-line file. Split into types.ts (106), catalog.ts (412, the
static PROVIDERS table + aliasing), model-overlay.ts (124, the runtime
overlay + getProvider/getProviderList). index.ts is now a 14-line barrel —
all 18+ external consumers import via "providers"/"./providers" unchanged.

tsc --noEmit clean; providers/index.test.ts (18 tests) and
pin-consistency.test.ts pass unchanged.

<!-- agentmux:agent_id=agentx -->
