+++ frontend/app/view/agent/providers/index.ts
@@ -1,633 +1,14 @@
 // Copyright 2025, AgentMux Corp.
 // SPDX-License-Identifier: Apache-2.0
 
-import { createSignal } from "solid-js";
+// Barrel: this module used to bundle everything (type contracts, the static
+// per-CLI catalog, and the dynamic model-catalog overlay) in one file. It's
+// now split for readability — see ./types, ./catalog, ./model-overlay — but
+// every external consumer keeps importing from "providers"/"providers/index"
+// unchanged.
 
-/**
- * A system tool the provider's CLI needs at runtime — beyond Node
- * itself, which is checked separately by `check_nodejs_available`.
- * Examples: Claude Code calls `git` from inside session-start; the
- * GitHub Copilot CLI wraps `gh`.
- *
- * Probed pre-launch via the `resolve_prereqs` RPC. Missing prereqs
- * open the `AgentPrereqModal` with platform-aware install links.
- * See docs/specs/SPEC_PROVIDER_SYSTEM_PREREQS_2026_05_18.md.
- */
-export interface SystemPrereq {
-    /** Binary name to look up via `where` (Windows) / `which` (Unix). */
-    tool: string;
-    /** Display label in the banner. Defaults to `tool` if omitted. */
-    label?: string;
-    /** Per-platform official install URLs. Curated landing pages so
-     *  the link reads as intentional, not a generic Google search. */
-    installUrls: {
-        windows: string;
-        macos: string;
-        linux: string;
-    };
-    /** Anchor text shown for the install link, per platform. Falls
-     *  back to "Install {label}" when omitted. */
-    installLinkText?: {
-        windows?: string;
-        macos?: string;
-        linux?: string;
-    };
-}
+export type { SystemPrereq, ProviderModel, ProviderDefinition } from "./types";
 
-/** One selectable model for a provider's `--model` flag. */
-export interface ProviderModel {
-    value: string;          // the model string passed to the CLI (e.g. "opus", "gpt-5.5")
-    label: string;          // UI label
-    default?: boolean;      // the default model for this provider
-    description?: string;   // optional one-liner shown in the /model picker
-    aliases?: string[];     // optional /model aliases (e.g. "claude-opus")
-}
+export { GIT_PREREQ, PROVIDERS, PROVIDER_ALIASES, resolveProviderAlias } from "./catalog";
 
-export interface ProviderDefinition {
-    id: string;
-    displayName: string;
-    cliCommand: string;
-    defaultArgs: string[];
-    styledArgs: string[];        // CLI flags for JSON streaming mode (documentation; use launchArgs for actual invocation)
-    outputFormat: "claude-stream-json" | "gemini-json" | "codex-json" | "kimi-stream-json" | "acp" | "raw";
-    styledOutputFormat: "claude-stream-json" | "gemini-json" | "codex-json" | "kimi-stream-json" | "acp";
-    authType: "oauth" | "api-key";
-    authCheckCommand: string[];  // e.g. ["auth", "status", "--json"]
-    authLoginCommand: string[];  // e.g. ["auth", "login"]
-    npmPackage: string;          // npm package name for local install
-    pinnedVersion: string;       // version to install ("latest" or specific)
-    docsUrl: string;
-    windowsInstallCommand: string;  // official installer for Windows (powershell)
-    unixInstallCommand: string;      // official installer for macOS/Linux (bash)
-    icon: string;
-    unsetEnv?: string[];         // env vars to unset before launching (e.g. nested-session guards)
-    // Auth isolation — each provider gets its own versioned config dir
-    authConfigDirEnvVar: string;        // env var that redirects the provider's config/auth dir
-    authDirName: string;                // subdir name under {dataDir}/auth/ (e.g. "claude")
-    authExtraEnv?: Record<string, string>;  // extra env vars needed for auth isolation (e.g. GEMINI_FORCE_FILE_STORAGE)
-    // Launch args — the complete CLI args for a single turn (replaces hardcoded ["-p", ...styledArgs])
-    // The user message is written to subprocess stdin; these args put the CLI in non-interactive mode.
-    launchArgs: string[];
-    // Resume flag — how to pass a session ID for multi-turn continuity.
-    // null means this provider does not support simple-flag resume (e.g. Codex uses a subcommand).
-    resumeFlag: string | null;
-    // JSON field name containing the session/thread ID in the CLI's init event.
-    sessionIdField: string;
-    // Controller type: "persistent" keeps a long-running process with stdin streaming,
-    // "subprocess" spawns a fresh process per turn with --resume,
-    // "acp" uses the Agent Client Protocol (JSON-RPC 2.0 over stdio).
-    controllerType: "persistent" | "subprocess" | "acp";
-    // Launch args for persistent mode (--input-format stream-json, no -p).
-    // Only used when controllerType is "persistent".
-    persistentLaunchArgs?: string[];
-    // Spawn the auth login subprocess under a PTY (instead of plain
-    // piped stdio). Required by providers whose auth subcommand checks
-    // `isatty()` and refuses to run otherwise — currently OpenClaw's
-    // `openclaw models auth login --provider <id>`. The host's
-    // `run_cli_login` reads this flag and chooses the PTY branch.
-    requiresLoginTty?: boolean;
-    /** System tools the provider's CLI needs at runtime. Probed before
-     *  the user launches the agent so we can show install links
-     *  instead of letting the CLI fail with cryptic stderr. */
-    systemPrereqs?: SystemPrereq[];
-    /**
-     * The model's maximum input token capacity (context window size).
-     * Used by the composer strip to render a context-fill progress bar.
-     * Omit for providers whose context window is unknown or variable.
-     */
-    contextWindow?: number;
-    /**
-     * AgentMux-side model choices for this provider's `--model` flag. Drives the
-     * `/model` slash command (and, for Claude, the control-bar dropdown). Mark one
-     * `default`. Omit for providers whose model is chosen in their own config
-     * (muxcode/openclaw/pi/copilot) — those show no AgentMux model picker.
-     * NOTE: model strings move with the upstream provider — keep current; see
-     * docs/providers/PROVIDER_MODELS_EFFORT_SETTINGS_2026-06.md.
-     */
-    models?: ProviderModel[];
-}
-
-/** Shared git prereq — claude-code calls `git` from session-start
- *  (issue anthropics/claude-code#29898), and openclaw uses git for
- *  project context. Curated install URLs per platform. */
-const GIT_PREREQ: SystemPrereq = {
-    tool: "git",
-    label: "Git",
-    installUrls: {
-        windows: "https://git-scm.com/download/win",
-        macos: "https://git-scm.com/download/mac",
-        linux: "https://git-scm.com/download/linux",
-    },
-    installLinkText: {
-        windows: "Install Git for Windows",
-        macos: "Install Git for macOS",
-        linux: "Install Git",
-    },
-};
-
-export const PROVIDERS: Record<string, ProviderDefinition> = {
-    claude: {
-        id: "claude",
-        displayName: "Claude Code",
-        cliCommand: "claude",
-        defaultArgs: [],
-        styledArgs: ["--output-format", "stream-json", "--verbose", "--include-partial-messages", "--dangerously-skip-permissions"],
-        outputFormat: "raw",
-        styledOutputFormat: "claude-stream-json",
-        authType: "oauth",
-        authCheckCommand: ["auth", "status", "--json"],
-        // NOTE (2026-06-23, SPEC_HOST_CLI_LOGIN_CAPTURE §0): in-app OAuth is a DEAD
-        // END for Claude v2.1.x — the login is a self-driving TUI that opens its own
-        // browser, and when WE spawn it (host→PTY, not a real terminal) the browser
-        // never opens (it hangs). `setup-token` shares the same browser front-end, so
-        // it can't help in-app either (confirmed via two live captures). The robust
-        // path is seed-from-global (§5.5): authenticate once in a real terminal, then
-        // seed. Kept as ["auth","login"] only for the legacy Connect CTA.
-        authLoginCommand: ["auth", "login"],
-        // Run `claude auth login` under a PTY (run_cli_login's PTY branch).
-        // Spawned with plain pipes from the GUI host, the CLI exits cleanly
-        // ~5s after printing the OAuth URL — before the user can return from
-        // the browser and paste the code — so the login always appeared stuck
-        // on the URL. (Reproduced extensively: standalone runs with a
-        // controlling terminal stay alive at the "Paste code here >" prompt;
-        // only the host's terminal-less pipe spawn exits early.) A PTY makes
-        // the CLI fully interactive (isTTY + a controlling terminal), so it
-        // stays alive for the paste. See docs / run_cli_login_pty.
-        requiresLoginTty: true,
-        npmPackage: "@anthropic-ai/claude-code",
-        // Keep in sync with agentmux-srv/src/backend/providers.rs `pinned_version`,
-        // agentmux-cef/src/commands/providers.rs `CLAUDE_VERSION`, and
-        // .github/workflows/container-image.yml `claude_version` default — enforced by
-        // ./pin-consistency.test.ts.
-        pinnedVersion: "2.1.198",
-        docsUrl: "https://docs.anthropic.com/claude-code",
-        windowsInstallCommand: "irm https://claude.ai/install.ps1 | iex",
-        unixInstallCommand: "curl -fsSL https://claude.ai/install.sh | bash",
-        icon: "sparkles",
-        unsetEnv: ["CLAUDECODE"],
-        authConfigDirEnvVar: "CLAUDE_CONFIG_DIR",
-        authDirName: "claude",
-        launchArgs: ["-p", "--output-format", "stream-json", "--verbose", "--include-partial-messages", "--dangerously-skip-permissions"],
-        resumeFlag: "--resume",
-        sessionIdField: "session_id",
-        // Persistent (bidirectional stream-json) + the Agent SDK CONTROL PROTOCOL
-        // is the only way AskUserQuestion works headless: the CLI auto-rejects it
-        // ("Error: Answer questions?") unless launched with `--permission-prompt-tool
-        // stdio` and answered via a control_response. `--dangerously-skip-permissions`
-        // DISABLES that routing, so it must NOT be here. The persistent controller's
-        // ControlChannel auto-allows ordinary tools to preserve today's yolo UX and
-        // surfaces only AskUserQuestion to the user. Keep in sync with `static CLAUDE`
-        // in agentmux-srv/providers.rs. controllerType selects persistentLaunchArgs
-        // over launchArgs in useAgentCommands.ts.
-        // Spec: docs/specs/SPEC_AGENT_CONTROL_PROTOCOL_2026_06_15.md.
-        controllerType: "persistent",
-        persistentLaunchArgs: ["--input-format", "stream-json", "--output-format", "stream-json", "--verbose", "--include-partial-messages", "--permission-prompt-tool", "stdio", "--permission-mode", "default"],
-        // Claude Code calls `git` at session-start (issue
-        // anthropics/claude-code#29898). Without git the CLI fails
-        // with `Error: Git is required but was not found.`.
-        systemPrereqs: [GIT_PREREQ],
-        contextWindow: 200_000,
-        // Labels carry the concrete version the pinned CLI (see `pinnedVersion`)
-        // currently resolves each family alias to — curated, kept in sync on a
-        // pin bump (SPEC_AGENT_MODEL_DROPDOWN_CLI_PIN_LOG Part B). The `value`
-        // stays the family alias (opus/sonnet/haiku), which `--model` resolves
-        // to exactly that version; when a family ever has TWO live versions,
-        // switch those entries to concrete `--model` IDs (e.g. claude-sonnet-4-6).
-        // `default: true` marks the family the runtime falls back to when no
-        // model is set — it MUST match DEFAULT_RUNTIME_CONFIG.model (types.ts),
-        // which is `sonnet` (faster default for routine turns). A mismatch here
-        // desyncs the strip's Model dropdown from the actual default and is a
-        // trap for any `models.find(m => m.default)` reader.
-        models: [
-            { value: "opus", label: "Opus 4.8", description: "Claude Opus 4.8 — highest quality", aliases: ["claude-opus"] },
-            { value: "sonnet", label: "Sonnet 5", default: true, description: "Claude Sonnet 5 — balanced", aliases: ["claude-sonnet"] },
-            { value: "haiku", label: "Haiku 4.5", description: "Claude Haiku 4.5 — fastest", aliases: ["claude-haiku"] },
-            // No confirmed generic "fable" alias (unlike opus/sonnet/haiku above), so this
-            // pins the concrete model id directly — same id already relied on by
-            // context-window.ts's 1M-context-window band. Kept current here so the
-            // offline/API-failure fallback still shows it (see setProviderModels below
-            // for how this stays in sync with the live catalog once reachable).
-            { value: "claude-fable-5", label: "Fable 5", description: "Claude Fable 5" },
-        ],
-    },
-    codex: {
-        id: "codex",
-        displayName: "Codex CLI",
-        cliCommand: "codex",
-        defaultArgs: [],
-        // exec subcommand runs non-interactively; --json emits NDJSON events; - reads prompt from stdin
-        styledArgs: ["exec", "--json", "--dangerously-bypass-approvals-and-sandbox", "-"],
-        outputFormat: "raw",
-        styledOutputFormat: "codex-json",
-        authType: "oauth",
-        authCheckCommand: ["login", "status"],
-        authLoginCommand: ["login"],
-        npmPackage: "@openai/codex",
-        pinnedVersion: "0.116.0",
-        docsUrl: "https://platform.openai.com/docs/codex",
-        windowsInstallCommand: "npm install -g @openai/codex",
-        unixInstallCommand: "npm install -g @openai/codex",
-        icon: "robot",
-        authConfigDirEnvVar: "CODEX_HOME",
-        authDirName: "codex",
-        launchArgs: ["exec", "--json", "--dangerously-bypass-approvals-and-sandbox", "-"],
-        // Codex resume requires a subcommand change (exec resume <id>), not a simple flag.
-        // Multi-turn is handled by re-running exec; null disables automatic --resume append.
-        resumeFlag: null,
-        sessionIdField: "thread_id",
-        controllerType: "subprocess",
-        contextWindow: 200_000,
-        // Verify ChatGPT-account availability when bumping the codex CLI pin.
-        models: [
-            { value: "gpt-5.5", label: "GPT-5.5", default: true, description: "Current codex frontier" },
-            { value: "gpt-5.4", label: "GPT-5.4", description: "Prior frontier" },
-            { value: "gpt-5.1-codex-max", label: "GPT-5.1-Codex-Max", description: "Codex-tuned, long-horizon" },
-            { value: "gpt-5.3-codex", label: "GPT-5.3-Codex", description: "Codex-tuned" },
-        ],
-    },
-    // muxcode — AgentMux's first-party agentic coding CLI.
-    // Supports local GGUF inference via llama-server, Anthropic, OpenAI, and
-    // OpenAI-compatible backends. Emits claude-stream-json NDJSON output.
-    // npm: @agentmuxai/muxcode
-    muxcode: {
-        id: "muxcode",
-        displayName: "Mux Code",
-        cliCommand: "muxcode",
-        defaultArgs: [],
-        // muxcode emits NDJSON unconditionally — no --output-format flag.
-        styledArgs: ["run", "-p"],
-        outputFormat: "raw",
-        styledOutputFormat: "claude-stream-json",
-        authType: "api-key",
-        // `auth status` exits 0 when any backend is ready (API key env var or
-        // local GGUF model installed), exits 1 when nothing is configured.
-        // Avoids the false-positive that `--version` would cause.
-        authCheckCommand: ["auth", "status"],
-        // `auth login` pulls a default local model when no backend is configured.
-        authLoginCommand: ["auth", "login"],
-        npmPackage: "@agentmuxai/muxcode",
-        pinnedVersion: "0.1.0",
-        docsUrl: "https://github.com/agentmuxai/muxcode",
-        windowsInstallCommand: "npm install -g @agentmuxai/muxcode",
-        unixInstallCommand: "npm install -g @agentmuxai/muxcode",
-        icon: "brain",
-        authConfigDirEnvVar: "MUXCODE_CONFIG_DIR",
-        authDirName: "muxcode",
-        launchArgs: ["run", "-p"],
-        resumeFlag: "--resume",
-        sessionIdField: "session_id",
-        controllerType: "subprocess",
-        contextWindow: 200_000,
-    },
-    gemini: {
-        id: "gemini",
-        displayName: "Gemini CLI",
-        cliCommand: "gemini",
-        defaultArgs: [],
-        // --output-format stream-json: NDJSON events; --yolo: auto-approve all tools;
-        // -p "": enable headless/non-interactive mode (prompt comes from stdin)
-        styledArgs: ["--output-format", "stream-json", "--yolo", "-p", ""],
-        outputFormat: "raw",
-        styledOutputFormat: "gemini-json",
-        authType: "oauth",
-        authCheckCommand: ["auth", "status"],
-        authLoginCommand: ["auth", "login"],
-        npmPackage: "@google/gemini-cli",
-        pinnedVersion: "0.32.1",
-        docsUrl: "https://ai.google.dev/gemini-cli",
-        windowsInstallCommand: "npm install -g @google/gemini-cli",
-        unixInstallCommand: "npm install -g @google/gemini-cli",
-        icon: "diamond",
-        authConfigDirEnvVar: "GEMINI_CLI_HOME",
-        authDirName: "gemini",
-        authExtraEnv: { GEMINI_FORCE_FILE_STORAGE: "true" },
-        launchArgs: ["--output-format", "stream-json", "--yolo", "-p", ""],
-        resumeFlag: "-r",
-        sessionIdField: "session_id",
-        controllerType: "subprocess",
-        contextWindow: 1_000_000,
-    },
-    // Qwen Code — Alibaba's open-source coding agent, a fork of Gemini CLI.
-    // Same stream-json headless surface → reuses the gemini translator
-    // (styledOutputFormat "gemini-json"). Backend is OpenAI-compatible: set
-    // OPENAI_BASE_URL=https://openrouter.ai/api/v1 + OPENAI_API_KEY (+ OPENAI_MODEL)
-    // to run any OpenRouter model. The Qwen OAuth free tier was retired
-    // (2026-04-15), so this is treated as api-key.
-    // Auth: the intended path is an env-injected key from the identity bundle
-    // (OPENAI_API_KEY/OPENROUTER_API_KEY), not an interactive CLI login. We use
-    // the Gemini-parent `auth status`/`auth` convention for the check/login:
-    // `auth status` fails-closed (prompts for auth) rather than reporting a
-    // false positive the way `--version` would (checkcliauth treats any
-    // non-JSON zero exit as authenticated — cli_handlers.rs). A deeper fix
-    // would validate the bound env key directly in the api-key flow.
-    qwen: {
-        id: "qwen",
-        displayName: "Qwen Code",
-        cliCommand: "qwen",
-        defaultArgs: [],
-        styledArgs: ["--output-format", "stream-json", "--yolo", "-p", ""],
-        outputFormat: "raw",
-        styledOutputFormat: "gemini-json",
-        authType: "api-key",
-        authCheckCommand: ["auth", "status"],
-        authLoginCommand: ["auth"],
-        npmPackage: "@qwen-code/qwen-code",
-        pinnedVersion: "0.19.2",
-        docsUrl: "https://qwenlm.github.io/qwen-code-docs",
-        windowsInstallCommand: "npm install -g @qwen-code/qwen-code",
-        unixInstallCommand: "npm install -g @qwen-code/qwen-code",
-        icon: "feather",
-        authConfigDirEnvVar: "QWEN_HOME",
-        authDirName: "qwen",
-        launchArgs: ["--output-format", "stream-json", "--yolo", "-p", ""],
-        resumeFlag: null,
-        sessionIdField: "session_id",
-        controllerType: "subprocess",
-    },
-    // OpenClaw — model-agnostic personal AI assistant from openclaw.ai.
-    // We launch its `openclaw acp` bridge: speaks ACP over stdio (our
-    // side) and forwards turns to OpenClaw's local Gateway daemon over
-    // WebSocket (its side). The Gateway is OpenClaw's own daemon
-    // (`openclaw gateway --port 18789`) and MUST be running before
-    // the bridge can establish a session — onboarding via
-    // `openclaw onboard` covers that on first install.
-    //
-    // OpenClaw is model-agnostic — the backing LLM brain is selected
-    // by the user inside OpenClaw's own config (defaults to Pi; users
-    // can wire Claude, Codex/OpenAI, Gemini, or local models via
-    // `openclaw models auth login --provider <provider>`).
-    openclaw: {
-        id: "openclaw",
-        displayName: "OpenClaw",
-        cliCommand: "openclaw",
-        defaultArgs: [],
-        styledArgs: [],
-        outputFormat: "acp",
-        styledOutputFormat: "acp",
-        // OAuth-via-subcommand. `openclaw models auth login --provider
-        // openai-codex` runs OpenAI's "Sign in with ChatGPT" flow and
-        // writes the resulting profile under ~/.openclaw/. OpenClaw then
-        // uses that profile to spawn Codex's app-server as the agent's
-        // backing brain (per docs.openclaw.ai/plugins/codex-harness +
-        // SPEC_OPENCLAW_AGENT_2026_05_17.md §4).
-        //
-        // Future Phase: add more provider options ("Login with Claude",
-        // "Login with Gemini", ...) and let the user pick which brain
-        // before the OAuth subcommand runs.
-        authType: "oauth",
-        // `doctor` is a health/repair command — exits 0 even when no
-        // openai-codex auth profile is registered, which would let
-        // AgentMux skip the OAuth login and launch `openclaw acp`
-        // unauthenticated. List the profiles for the specific provider
-        // instead; exits non-zero when none are configured.
-        authCheckCommand: ["models", "auth", "list", "--provider", "openai-codex"],
-        // `cliCommand: "openclaw"` is prefixed by the spawn layer — keep
-        // only the args here. Kimi's `["login"]` is the convention.
-        authLoginCommand: ["models", "auth", "login", "--provider", "openai-codex"],
-        npmPackage: "openclaw",
-        pinnedVersion: "2026.6.10",
-        docsUrl: "https://docs.openclaw.ai",
-        windowsInstallCommand: "npm install -g openclaw",
-        unixInstallCommand: "npm install -g openclaw",
-        icon: "lobster",
-        authConfigDirEnvVar: "OPENCLAW_HOME",
-        authDirName: "openclaw",
-        launchArgs: ["acp"],
-        resumeFlag: null,
-        sessionIdField: "sessionId",
-        controllerType: "acp",
-        requiresLoginTty: true,
-        // Same git dependency as Claude Code — OpenClaw uses git for
-        // project-context features when invoking the Codex harness.
-        systemPrereqs: [GIT_PREREQ],
-        contextWindow: 200_000,
-    },
-    // Kimi Code CLI — Moonshot AI's coding agent.
-    // Python-based CLI (not npm). Supports stream-json output and OpenAI-style tool calls.
-    kimi: {
-        id: "kimi",
-        displayName: "Kimi Code CLI",
-        cliCommand: "kimi",
-        defaultArgs: [],
-        styledArgs: ["--print", "--output-format", "stream-json", "--yolo", "-p", ""],
-        outputFormat: "raw",
-        styledOutputFormat: "kimi-stream-json",
-        authType: "api-key",
-        authCheckCommand: ["info"],
-        authLoginCommand: ["login"],
-        npmPackage: "",
-        pinnedVersion: "",
-        docsUrl: "https://moonshotai.github.io/kimi-cli/",
-        windowsInstallCommand: "pip install kimi-cli",
-        unixInstallCommand: "pip install kimi-cli",
-        icon: "moon",
-        unsetEnv: [],
-        authConfigDirEnvVar: "KIMI_SHARE_DIR",
-        authDirName: "kimi",
-        launchArgs: ["--print", "--output-format", "stream-json", "--yolo", "-p", ""],
-        resumeFlag: null,
-        sessionIdField: "session_id",
-        controllerType: "subprocess",
-        contextWindow: 128_000,
-    },
-    // GitHub Copilot CLI — Microsoft's coding agent.
-    // Runs in ACP mode (`--acp` flag) so the existing ACP controller
-    // can drive it the same way it drives Pi and OpenClaw. The CLI's
-    // `-p`/`--prompt` non-interactive mode doesn't accept stdin for
-    // the prompt yet (github/copilot-cli#96, #1046), so ACP is the
-    // only path that composes with our existing subprocess+stdin
-    // controller model. Documentation and CLI reference come from
-    // discussion #493 (research-cli-context-files-2026-04-22).
-    copilot: {
-        id: "copilot",
-        displayName: "GitHub Copilot CLI",
-        cliCommand: "copilot",
-        defaultArgs: [],
-        styledArgs: ["--acp"],
-        outputFormat: "acp",
-        styledOutputFormat: "acp",
-        authType: "oauth",
-        authCheckCommand: ["auth", "status"],
-        authLoginCommand: ["auth", "login"],
-        npmPackage: "@github/copilot",
-        pinnedVersion: "1.0.65",
-        docsUrl: "https://docs.github.com/copilot/concepts/agents/about-copilot-cli",
-        windowsInstallCommand: "npm install -g @github/copilot",
-        unixInstallCommand: "npm install -g @github/copilot",
-        icon: "github",
-        authConfigDirEnvVar: "COPILOT_HOME",
-        authDirName: "copilot",
-        launchArgs: ["--acp"],
-        resumeFlag: null,
-        sessionIdField: "sessionId",
-        controllerType: "acp",
-        contextWindow: 128_000,
-    },
-    // Pi — the lightweight coding agent that powers OpenClaw.
-    // Standalone CLI, no gateway required. Pure coding agent with read/write/bash/edit tools.
-    // Ideal when users want a fast, self-contained coding agent without the full OpenClaw stack.
-    pi: {
-        id: "pi",
-        displayName: "Pi",
-        cliCommand: "pi",
-        defaultArgs: [],
-        styledArgs: ["--json"],
-        outputFormat: "acp",
-        styledOutputFormat: "acp",
-        authType: "api-key",
-        authCheckCommand: ["config", "get", "provider"],
-        authLoginCommand: ["config"],
-        npmPackage: "@mariozechner/pi-coding-agent",
-        pinnedVersion: "0.73.1",
-        docsUrl: "https://github.com/badlogic/pi-mono",
-        windowsInstallCommand: "npm install -g @mariozechner/pi-coding-agent",
-        unixInstallCommand: "npm install -g @mariozechner/pi-coding-agent",
-        icon: "terminal",
-        authConfigDirEnvVar: "PI_HOME",
-        authDirName: "pi",
-        launchArgs: ["--json"],
-        resumeFlag: null,
-        sessionIdField: "sessionId",
-        controllerType: "acp",
-    },
-};
-
-// Aliases for provider IDs from older databases or alternate naming
-const PROVIDER_ALIASES: Record<string, string> = {
-    "claude-code": "claude",
-    "claude_code": "claude",
-    "codex-cli": "codex",
-    "gemini-cli": "gemini",
-    "qwen-code": "qwen",
-    "qwen3-coder": "qwen",
-    "kimi-cli": "kimi",
-    "kimi_code": "kimi",
-    "openclaw-cli": "openclaw",
-    "open-claw": "openclaw",
-    "copilot-cli": "copilot",
-    "github-copilot": "copilot",
-    "copilot_cli": "copilot",
-    "mux-code": "muxcode",
-    "mux_code": "muxcode",
-};
-
-export function resolveProviderAlias(id: string): string {
-    return PROVIDER_ALIASES[id] ?? id;
-}
-
-// ── API-sourced model catalog overlay ───────────────────────────────────────
-//
-// The static `models` above are hand-curated and their version labels drift
-// ("Sonnet 4.6" → "Sonnet 5" …). `providers.models` (backend) reads the
-// authoritative list from the Anthropic Models API; `setProviderModels` folds
-// it in. Signal-backed so a strip that rendered before the fetch resolved
-// re-renders when the catalog lands (see app-init.ts).
-const [modelOverlay, setModelOverlay] = createSignal<Record<string, ProviderModel[]>>({});
-
-/** Numeric version tokens of a model id, e.g. "claude-sonnet-5" → [5],
- *  "claude-opus-4-8" → [4,8]. Order-independent "which is newest" key. */
-function versionTuple(id: string): number[] {
-    return (id.match(/\d+/g) ?? []).map((n) => parseInt(n, 10));
-}
-
-/** Element-wise version compare; missing positions count as 0 so [5] > [3,5]. */
-function cmpVersion(a: number[], b: number[]): number {
-    const n = Math.max(a.length, b.length);
-    for (let i = 0; i < n; i++) {
-        const d = (a[i] ?? 0) - (b[i] ?? 0);
-        if (d !== 0) return d;
-    }
-    return 0;
-}
-
-/** The highest-versioned model in a group — robust to API ordering, so a newly
- *  shipped Sonnet 5.x / 6 wins over the prior version automatically. */
-function pickNewest(models: ProviderModel[]): ProviderModel {
-    return models.reduce((best, m) =>
-        cmpVersion(versionTuple(m.value), versionTuple(best.value)) > 0 ? m : best,
-    );
-}
-
-/** Family key for an id the curated list doesn't cover — its alphabetic tokens
- *  minus the "claude" prefix, e.g. "claude-fable-5" → "fable". Groups versions
- *  of a genuinely new family so we surface one (newest) entry for it. */
-function familyKey(id: string): string {
-    const alpha = id.split("-").filter((t) => /^[a-z]+$/i.test(t) && t.toLowerCase() !== "claude");
-    return alpha.join("-").toLowerCase() || id.toLowerCase();
-}
-
-/** Drop the redundant "Claude " prefix so API labels match the curated style
- *  ("Claude Sonnet 5" → "Sonnet 5"). */
-function cleanLabel(label: string): string {
-    return label.replace(/^claude\s+/i, "").trim();
-}
-
-/**
- * Fold the authoritative catalog into a provider's model list. Behavior-
- * preserving by design:
- *  - Curated family entries (opus/sonnet/haiku/fable) keep their curated
- *    `value` (for opus/sonnet/haiku, a short alias `--model` resolves to the
- *    current version; fable has no such alias yet, so its `value` is the
- *    concrete pinned id instead), `default` marker, `aliases`, and
- *    `description`; only their **label** is refreshed to the newest matching
- *    API model. This is what turns "Sonnet 4.6" into "Sonnet 5" without changing
- *    what `--model` receives or breaking `models.find(m => m.default)`.
- *  - Families the curated list doesn't cover (any future family) are appended
- *    automatically, one newest entry each, using the concrete API id as
- *    `value` (always a valid `--model` target). No family is hardcoded.
- *  - Matching uses `familyKey()` (alphabetic tokens, "claude" stripped, digits
- *    dropped) rather than a raw substring test, so a *version-pinned* curated
- *    `value` (fable's `claude-fable-5`) still matches a future API version
- *    (`claude-fable-6`) the same way a generic alias (`sonnet`) already does —
- *    a plain `.includes()` test would only match the exact pinned string,
- *    leaving the curated row stale AND appending a duplicate "extra" entry
- *    once a newer version showed up in the API.
- * Empty input (no token / macOS Keychain / offline) is a no-op → static list.
- */
-export function setProviderModels(id: string, apiModels: ProviderModel[]): void {
-    if (apiModels.length === 0) return;
-    const canonical = resolveProviderAlias(id);
-    const base = PROVIDERS[id] ?? PROVIDERS[canonical];
-    if (!base) return;
-
-    const consumed = new Set<string>();
-
-    // 1. Refresh curated family labels from the newest API model in that family.
-    const curated = base.models.map((m) => {
-        const family = familyKey(m.value);
-        const matches = apiModels.filter((a) => familyKey(a.value) === family);
-        if (matches.length === 0) return m;
-        matches.forEach((a) => consumed.add(a.value)); // don't re-surface as an "extra"
-        return { ...m, label: cleanLabel(pickNewest(matches).label) };
-    });
-
-    // 2. Surface families the curated list misses (grouped, newest per family).
-    const byFamily = new Map<string, ProviderModel[]>();
-    for (const a of apiModels) {
-        if (consumed.has(a.value)) continue;
-        const key = familyKey(a.value);
-        const group = byFamily.get(key) ?? [];
-        group.push(a);
-        byFamily.set(key, group);
-    }
-    const extras = [...byFamily.values()].map((group) => {
-        const newest = pickNewest(group);
-        return { value: newest.value, label: cleanLabel(newest.label) } satisfies ProviderModel;
-    });
-
-    setModelOverlay((prev) => ({ ...prev, [canonical]: [...curated, ...extras] }));
-}
-
-export function getProvider(id: string): ProviderDefinition | undefined {
-    const base = PROVIDERS[id] ?? PROVIDERS[resolveProviderAlias(id)];
-    if (!base) return undefined;
-    const overlay = modelOverlay()[resolveProviderAlias(base.id)];
-    return overlay ? { ...base, models: overlay } : base;
-}
-
-export function getProviderList(): ProviderDefinition[] {
-    // Route through getProvider so list consumers that read `.models` see the
-    // same API-sourced overlay (refreshed labels / new families) as callers of
-    // getProvider(), not the raw static list.
-    return Object.keys(PROVIDERS).map((id) => getProvider(id)!);
-}
+export { setProviderModels, getProvider, getProviderList } from "./model-overlay";
